Fix selecting coordinates from map
